# schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "current": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "SetupIn": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "title": "Password",
            "type": "string"
          },
          "timezone": {
            "default": "UTC",
            "title": "Timezone",
            "type": "string"
          },
          "workspace_name": {
            "default": "My workspace",
            "maxLength": 200,
            "minLength": 1,
            "title": "Workspace Name",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "SetupIn",
        "type": "object"
      },
      "SetupOut": {
        "properties": {
          "otpauth_uri": {
            "title": "Otpauth Uri",
            "type": "string"
          },
          "setup_token": {
            "title": "Setup Token",
            "type": "string"
          }
        },
        "required": [
          "otpauth_uri",
          "setup_token"
        ],
        "title": "SetupOut",
        "type": "object"
      },
...
      "UsageRow": {
        "properties": {
          "counter": {
            "title": "Counter",
            "type": "string"
          },
          "day": {
            "format": "date",
            "title": "Day",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "day",
          "counter",
          "value"
        ],
        "title": "UsageRow",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Tumnis Guide REST API. Every error is `application/problem+json` with a stable\n`code`. Every write takes an `Idempotency-Key` header (8 to 255 of `A-Z a-z 0-9 _ - : .`):\na retry with the same key within 24 hours replays the stored response and adds the header\n`Idempotent-Replayed: true`; the same key with a different request is 422\n`idempotency_mismatch`. Writes name the `version` they read; a stale one is 409\n`stale_version` with `current`. Lists page by `cursor` and `limit` and return\n`next_cursor`. Limits: 429 `rate_limited` with `Retry-After`, 413 `body_too_large`.",
    "title": "Tumnis Guide",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/setup": {
      "post": {
        "description": "First run: creates the workspace and its owner and shows the TOTP secret once.\n409 `already_set_up`; 403 `signup_disabled` in hosted mode.",
        "operationId": "auth_setup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad request (`idempotency_key_required`, `invalid_cursor`, ...)"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthenticated (`unauthenticated`, `session_expired`)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden (`csrf_failed`, `bad_origin`, `session_required`, ...)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict (`stale_version` with `current`)"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Body too large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error or `idempotency_mismatch`"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited (`Retry-After`)"
          }
        },
        "summary": "Setup",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/usage": {
      "get": {
        "operationId": "usage_get_usage",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsageRow"
                  },
                  "title": "Response Usage Get Usage",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad request (`idempotency_key_required`, `invalid_cursor`, ...)"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthenticated (`unauthenticated`, `session_expired`)"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden (`csrf_failed`, `bad_origin`, `session_required`, ...)"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict (`stale_version` with `current`)"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Body too large"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error or `idempotency_mismatch`"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited (`Retry-After`)"
          }
        },
        "summary": "Get Usage",
        "tags": [
          "usage"
        ]
      }
    }
  }
}